+++ build.gradle
@@ -64,14 +64,14 @@ sourceSets {
     main.java.srcDir querydslDir
 }
 
-//tasks.register("compileQuerydsl", JavaCompile) {
-//    group = "build"
-//    description = "Compiles the Querydsl Q-types"
-//    classpath = sourceSets.main.compileClasspath
-//    destinationDirectory.set(file(querydslDir))
-//    source = sourceSets.main.java
-//    options.annotationProcessorPath = configurations.annotationProcessor
-//}
+tasks.register("compileQuerydsl", JavaCompile) {
+    group = "build"
+    description = "Compiles the Querydsl Q-types"
+    classpath = sourceSets.main.compileClasspath
+    destinationDirectory.set(file(querydslDir))
+    source = sourceSets.main.java
+    options.annotationProcessorPath = configurations.annotationProcessor
+}
 
 tasks.named('test') {
 	useJUnitPlatform()
