+++ src/main/java/com/example/dgu_semi_erp_back/usecase/announcement/AnnouncementUseCase.java
@@ -5,6 +5,4 @@ import com.example.dgu_semi_erp_back.entity.announcement.Announcement;
 // Detail 조회때 쓰는 명세
 public interface AnnouncementUseCase {
     Announcement findAnnouncementById(Long id);
-
-    void deleteAnnouncement(Long id);
 }
