+++ src/main/java/com/example/dgu_semi_erp_back/usecase/announcement/DeleteAnnouncementUsecase.java
@@ -1,0 +1,5 @@
+package com.example.dgu_semi_erp_back.usecase.announcement;
+
+public interface DeleteAnnouncementUsecase {
+    void deleteAnnouncement(Long id);
+}
