+++ src/main/java/com/example/dgu_semi_erp_back/usecase/announcement/FindAnnouncementSummariesUseCase.java
@@ -6,6 +6,7 @@ import org.springframework.data.domain.Pageable;
 
 
 import java.time.LocalDate;
+import java.time.LocalDateTime;
 
 // 목록 조회시 쓰는 명세
 public interface FindAnnouncementSummariesUseCase {
